Adjust modules
