+++ cocos/core/animation/animation-clip.ts
@@ -36,23 +36,25 @@ import { DataPoolManager } from '../../3d/skeletal-animation/data-pool-manager';
 import { binarySearchEpsilon } from '../algorithm/binary-search';
 import { murmurhash2_32_gc } from '../utils/murmurhash2_gc';
 import { SkelAnimDataHub } from '../../3d/skeletal-animation/skeletal-animation-data-hub';
-import { ComponentPath, HierarchyPath, TargetPath, evaluatePath, isPropertyPath } from './target-path';
+import { ComponentPath, evaluatePath, isPropertyPath } from './target-path';
 import { WrapMode as AnimationWrapMode, WrapMode, WrapModeMask } from './types';
 import { IValueProxyFactory } from './value-proxy';
 import { legacyCC } from '../global-exports';
-import { RealCurve, RealInterpMode } from '../curves';
-import { ObjectCurve } from '../curves/object-curve';
-import { Color, Mat4, Quat, Size, Vec2, Vec3, Vec4 } from '../math';
+import { Mat4, Quat, Vec3 } from '../math';
 import { Node } from '../scene-graph/node';
-import { IntegerCurve } from '../curves/integer-curve';
-import { QuaternionCurve, QuaternionInterpMode } from '../curves/quat-curve';
-import { KeySharedQuaternionCurves, KeySharedRealCurves } from '../curves/keys-shared-curves';
 import { assertIsTrue } from '../data/utils/asserts';
 import type { PoseOutput } from './pose-output';
 import * as legacy from './legacy-clip-data';
 import { BAKE_SKELETON_CURVE_SYMBOL } from './internal-symbols';
-import { RealKeyframeValue } from '../curves/curve';
-import { CubicSplineNumberValue, CubicSplineVec2Value, CubicSplineVec3Value, CubicSplineVec4Value } from './cubic-spline-value';
+import { Binder, isTargetingTRS, RuntimeBinding, Track, TrackEval, TrackPath, TrsTrackPath } from './tracks/track';
+import { createEvalSymbol } from './define';
+import { VectorTrack } from './tracks/vector-track';
+import { UntypedTrack, UntypedTrackRefine } from './tracks/untyped-track';
+import { Range } from './tracks/utils';
+import { ObjectTrack } from './tracks/object-track';
+import { CompressedData, CompressedDataEvaluator } from './compression/compressed-data';
+import { RealTrack } from './tracks/real-track';
+import { QuaternionTrack } from './tracks/quat-track';
 
 export declare namespace AnimationClip {
     export interface IEvent {
@@ -66,428 +68,8 @@ export declare namespace AnimationClip {
 
 // #region Tracks
 
-type TrackPath = TargetPath[];
-
-interface Range {
-    min: number;
-    max: number;
-}
-
-const createEvalSymbol = Symbol('CreateEval');
-
-const CLASS_NAME_PREFIX_ANIM = 'cc.animation.';
-
 // Export for test
 export const searchForRootBonePathSymbol = Symbol('SearchForRootBonePath');
-
-/**
- * A track describes the path of animate a target.
- * It's the basic unit of animation clip.
- */
-@ccclass(`${CLASS_NAME_PREFIX_ANIM}Track`)
-export class Track {
-    @serializable
-    public path: TrackPath = [];
-
-    @serializable
-    public setter!: IValueProxyFactory | undefined;
-
-    public getChannels (): Channel[] {
-        return [];
-    }
-
-    public getRange (): Range {
-        const range: Range = { min: Infinity, max: -Infinity };
-        for (const channel of this.getChannels()) {
-            range.min = Math.min(range.min, channel.curve.rangeMin);
-            range.max = Math.max(range.max, channel.curve.rangeMax);
-        }
-        return range;
-    }
-
-    public [createEvalSymbol] (runtimeBinding: RuntimeBinding): TrackEval {
-        throw new Error(`No Impl`);
-    }
-}
-
-interface TrackEval {
-    /**
-     * Evaluates the track.
-     * @param time The time.
-     */
-    evaluate(time: number, runtimeBinding: RuntimeBinding): unknown;
-}
-
-type Curve = RealCurve | IntegerCurve | QuaternionCurve | ObjectCurve<unknown>;
-
-@ccclass(`${CLASS_NAME_PREFIX_ANIM}Channel`)
-export class Channel<T = Curve> {
-    constructor (curve: T) {
-        this._curve = curve;
-    }
-
-    @serializable
-    public name = '';
-
-    get curve () {
-        return this._curve;
-    }
-
-    @serializable
-    private _curve!: T;
-}
-
-type RealChannel = Channel<RealCurve>;
-
-type IntegerChannel = Channel<IntegerCurve>;
-
-type QuaternionChannel = Channel<QuaternionCurve>;
-
-@ccclass(`${CLASS_NAME_PREFIX_ANIM}SingleChannelTrack`)
-export abstract class SingleChannelTrack<TCurve extends Curve> extends Track {
-    constructor () {
-        super();
-        this._channel = new Channel<TCurve>(this.createCurve());
-    }
-
-    get channel () {
-        return this._channel;
-    }
-
-    public getChannels () {
-        return [this._channel];
-    }
-
-    protected createCurve (): TCurve {
-        throw new Error(`Not impl`);
-    }
-
-    public [createEvalSymbol] (_runtimeBinding: RuntimeBinding): TrackEval {
-        const { curve } = this._channel;
-        return {
-            evaluate: (time) => curve.evaluate(time),
-        };
-    }
-
-    @serializable
-    private _channel: Channel<TCurve>;
-}
-
-@ccclass(`${CLASS_NAME_PREFIX_ANIM}RealTrack`)
-export class RealTrack extends SingleChannelTrack<RealCurve> {
-    protected createCurve () {
-        return new RealCurve();
-    }
-}
-
-@ccclass(`${CLASS_NAME_PREFIX_ANIM}IntegerTrack`)
-export class IntegerTrack extends SingleChannelTrack<IntegerCurve> {
-    protected createCurve () {
-        return new IntegerCurve();
-    }
-}
-
-@ccclass(`${CLASS_NAME_PREFIX_ANIM}QuaternionTrack`)
-export class QuaternionTrack extends SingleChannelTrack<QuaternionCurve> {
-    protected createCurve () {
-        return new QuaternionCurve();
-    }
-
-    public [createEvalSymbol] () {
-        return new QuatTrackEval(this.getChannels()[0].curve);
-    }
-}
-
-class QuatTrackEval {
-    constructor (private _curve: QuaternionCurve) {
-
-    }
-
-    public evaluate (time: number) {
-        this._curve.evaluate(time, this._result);
-        return this._result;
-    }
-
-    private _result: Quat = new Quat();
-}
-
-@ccclass(`${CLASS_NAME_PREFIX_ANIM}ObjectTrack`)
-export class ObjectTrack<T> extends SingleChannelTrack<ObjectCurve<unknown>> {
-    protected createCurve () {
-        return new ObjectCurve<T>();
-    }
-}
-
-function maskIfEmpty<T extends Curve> (curve: T) {
-    return curve.empty ? undefined : curve;
-}
-
-@ccclass(`${CLASS_NAME_PREFIX_ANIM}VectorTrack`)
-export class VectorTrack extends Track {
-    constructor () {
-        super();
-        this._channels = new Array(4) as VectorTrack['_channels'];
-        for (let i = 0; i < this._channels.length; ++i) {
-            const channel = new Channel<RealCurve>(new RealCurve());
-            channel.name = 'X';
-            this._channels[i] = channel;
-        }
-    }
-
-    get componentsCount () {
-        return this._nComponents;
-    }
-
-    set componentsCount (value) {
-        this._nComponents = value;
-    }
-
-    public getChannels () {
-        return this._channels;
-    }
-
-    public [createEvalSymbol] () {
-        switch (this._nComponents) {
-        default:
-        case 2:
-            return new Vec2TrackEval(
-                maskIfEmpty(this._channels[0].curve),
-                maskIfEmpty(this._channels[1].curve),
-            );
-        case 3:
-            return new Vec3TrackEval(
-                maskIfEmpty(this._channels[0].curve),
-                maskIfEmpty(this._channels[1].curve),
-                maskIfEmpty(this._channels[2].curve),
-            );
-        case 4:
-            return new Vec4TrackEval(
-                maskIfEmpty(this._channels[0].curve),
-                maskIfEmpty(this._channels[1].curve),
-                maskIfEmpty(this._channels[2].curve),
-                maskIfEmpty(this._channels[3].curve),
-            );
-        }
-    }
-
-    @serializable
-    private _channels: [RealChannel, RealChannel, RealChannel, RealChannel];
-
-    @serializable
-    private _nComponents: 2 | 3 | 4 = 4;
-}
-
-class Vec2TrackEval {
-    constructor (private _x: RealCurve | undefined, private _y: RealCurve | undefined) {
-
-    }
-
-    public evaluate (time: number, runtimeBinding: RuntimeBinding) {
-        if ((!this._x || !this._y) && runtimeBinding.getValue) {
-            Vec2.copy(this._result, runtimeBinding.getValue() as Vec2);
-        }
-
-        if (this._x) {
-            this._result.x = this._x.evaluate(time);
-        }
-        if (this._y) {
-            this._result.y = this._y.evaluate(time);
-        }
-
-        return this._result;
-    }
-
-    private _result: Vec2 = new Vec2();
-}
-
-class Vec3TrackEval {
-    constructor (private _x: RealCurve | undefined, private _y: RealCurve | undefined, private _z: RealCurve | undefined) {
-
-    }
-
-    public evaluate (time: number, runtimeBinding: RuntimeBinding) {
-        if ((!this._x || !this._y || !this._z) && runtimeBinding.getValue) {
-            Vec3.copy(this._result, runtimeBinding.getValue() as Vec3);
-        }
-
-        if (this._x) {
-            this._result.x = this._x.evaluate(time);
-        }
-        if (this._y) {
-            this._result.y = this._y.evaluate(time);
-        }
-        if (this._z) {
-            this._result.z = this._z.evaluate(time);
-        }
-
-        return this._result;
-    }
-
-    private _result: Vec3 = new Vec3();
-}
-
-class Vec4TrackEval {
-    constructor (
-        private _x: RealCurve | undefined,
-        private _y: RealCurve | undefined,
-        private _z: RealCurve | undefined,
-        private _w: RealCurve | undefined,
-    ) {
-
-    }
-
-    public evaluate (time: number, runtimeBinding: RuntimeBinding) {
-        if ((!this._x || !this._y || !this._z || !this._w) && runtimeBinding.getValue) {
-            Vec4.copy(this._result, runtimeBinding.getValue() as Vec4);
-        }
-
-        if (this._x) {
-            this._result.x = this._x.evaluate(time);
-        }
-        if (this._y) {
-            this._result.y = this._y.evaluate(time);
-        }
-        if (this._z) {
-            this._result.z = this._z.evaluate(time);
-        }
-        if (this._w) {
-            this._result.w = this._w.evaluate(time);
-        }
-
-        return this._result;
-    }
-
-    private _result: Vec4 = new Vec4();
-}
-
-@ccclass(`${CLASS_NAME_PREFIX_ANIM}ColorTrack`)
-export class ColorTrack extends Track {
-    constructor () {
-        super();
-        this._channels = new Array(4) as ColorTrack['_channels'];
-        for (let i = 0; i < this._channels.length; ++i) {
-            const channel = new Channel<IntegerCurve>(new IntegerCurve());
-            channel.name = 'R';
-            this._channels[i] = channel;
-        }
-    }
-
-    public getChannels () {
-        return this._channels;
-    }
-
-    public [createEvalSymbol] () {
-        return new ColorTrackEval(
-            maskIfEmpty(this._channels[0].curve),
-            maskIfEmpty(this._channels[1].curve),
-            maskIfEmpty(this._channels[2].curve),
-            maskIfEmpty(this._channels[3].curve),
-        );
-    }
-
-    @serializable
-    private _channels: [IntegerChannel, IntegerChannel, IntegerChannel, IntegerChannel];
-}
-
-class ColorTrackEval<TCurve extends IntegerCurve | RealCurve = IntegerCurve> {
-    constructor (
-        private _x: TCurve | undefined,
-        private _y: TCurve | undefined,
-        private _z: TCurve | undefined,
-        private _w: TCurve | undefined,
-    ) {
-
-    }
-
-    public evaluate (time: number, runtimeBinding: RuntimeBinding) {
-        if ((!this._x || !this._y || !this._z || !this._w) && runtimeBinding.getValue) {
-            Color.copy(this._result, runtimeBinding.getValue() as Color);
-        }
-
-        if (this._x) {
-            this._result.r = this._x.evaluate(time);
-        }
-        if (this._y) {
-            this._result.g = this._y.evaluate(time);
-        }
-        if (this._z) {
-            this._result.b = this._z.evaluate(time);
-        }
-        if (this._w) {
-            this._result.a = this._w.evaluate(time);
-        }
-
-        return this._result;
-    }
-
-    private _result: Color = new Color();
-}
-
-@ccclass(`${CLASS_NAME_PREFIX_ANIM}UntypedTrackChannel`)
-class UntypedTrackChannel extends Channel<RealCurve> {
-    @serializable
-    public property!: string;
-
-    constructor () {
-        super(new RealCurve());
-    }
-}
-
-@ccclass(`${CLASS_NAME_PREFIX_ANIM}UntypedTrack`)
-class UntypedTrack extends Track {
-    @serializable
-    private _channels: UntypedTrackChannel[] = [];
-
-    public getChannels () {
-        return this._channels;
-    }
-
-    public [createEvalSymbol] (runtimeBinding: RuntimeBinding) {
-        if (!runtimeBinding.getValue) {
-            throw new Error(`Can not decide type for untyped track: runtime binding does not provide a getter.`);
-        }
-        const trySearchCurve = (property: string) => this._channels.find((channel) => channel.property === property)?.curve;
-        const value = runtimeBinding.getValue();
-        switch (true) {
-        case value instanceof Size:
-        default:
-            throw new Error(`Can not decide type for untyped track: got a unsupported value from runtime binding.`);
-        case value instanceof Vec2:
-            return new Vec2TrackEval(
-                trySearchCurve('x'),
-                trySearchCurve('y'),
-            );
-        case value instanceof Vec3:
-            return new Vec3TrackEval(
-                trySearchCurve('x'),
-                trySearchCurve('y'),
-                trySearchCurve('z'),
-            );
-        case value instanceof Vec4:
-            return new Vec4TrackEval(
-                trySearchCurve('x'),
-                trySearchCurve('y'),
-                trySearchCurve('z'),
-                trySearchCurve('w'),
-            );
-        case value instanceof Color:
-            // TODO: what if x, y, z, w?
-            return new ColorTrackEval(
-                trySearchCurve('r'),
-                trySearchCurve('g'),
-                trySearchCurve('b'),
-                trySearchCurve('a'),
-            );
-        }
-    }
-
-    public addChannel (property: string): UntypedTrackChannel {
-        const channel = new UntypedTrackChannel();
-        channel.property = property;
-        this._channels.push(channel);
-        return channel;
-    }
-}
 
 // #endregion
 
@@ -865,64 +447,15 @@ export class AnimationClip extends Asset {
      * @param refine How to decide the type on specified path.
      * @internal DO NOT USE THIS IN YOUR CODE.
      */
-    public upgradeUntypedTracks (refine: (path: TrackPath, setter?: IValueProxyFactory) => 'vec2' | 'vec3' | 'vec4' | 'color' | 'size') {
+    public upgradeUntypedTracks (refine: UntypedTrackRefine) {
         const newTracks: Track[] = [];
         for (const track of this._tracks) {
             if (!(track instanceof UntypedTrack)) {
                 continue;
             }
-            const untypedTrack = track;
-
-            const trySearchChannel = (property: string, outChannel: RealChannel) => {
-                const untypedChannel = untypedTrack.getChannels().find((channel) => channel.property === property);
-                if (untypedChannel) {
-                    outChannel.name = untypedChannel.name;
-                    outChannel.curve.assignSorted(
-                        Array.from(untypedChannel.curve.times()),
-                        Array.from(untypedChannel.curve.values()),
-                    );
-                }
-            };
-            const kind = refine(track.path, track.setter);
-            switch (kind) {
-            default:
-                continue;
-            case 'vec2': case 'vec3': case 'vec4': {
-                const track = new VectorTrack();
-                newTracks.push(track);
-                track.componentsCount = kind === 'vec2' ? 2 : kind === 'vec3' ? 3 : 4;
-                const [x, y, z, w] = track.getChannels();
-                switch (kind) {
-                case 'vec4':
-                    trySearchChannel('w', w);
-                    // fall through
-                case 'vec3':
-                    trySearchChannel('z', z);
-                    // fall through
-                default:
-                case 'vec2':
-                    trySearchChannel('x', x);
-                    trySearchChannel('y', y);
-                }
-                break;
-            }
-            case 'color': {
-                const track = new ColorTrack();
-                newTracks.push(track);
-                const [r, g, b, a] = track.getChannels();
-                trySearchChannel('r', r);
-                trySearchChannel('g', g);
-                trySearchChannel('b', b);
-                trySearchChannel('a', a);
-                // TODO: we need float-int conversion if xyzw
-                trySearchChannel('x', r);
-                trySearchChannel('y', g);
-                trySearchChannel('z', b);
-                trySearchChannel('w', a);
-                break;
-            }
-            case 'size':
-                break;
+            const newTrack = track.upgrade(refine);
+            if (newTrack) {
+                newTracks.push(newTrack);
             }
         }
     }
@@ -1242,214 +775,7 @@ export class AnimationClip extends Asset {
     }
 
     private _fromLegacy (legacyData: legacy.AnimationClipLegacyData) {
-        const newTracks: Track[] = [];
-
-        const {
-            keys: legacyKeys,
-            curves: legacyCurves,
-            commonTargets: legacyCommonTargets,
-        } = legacyData;
-
-        const untypedTracks = legacyCommonTargets.map((legacyCommonTarget) => {
-            const track = new UntypedTrack();
-            track.path = legacyCommonTarget.modifiers;
-            track.setter = legacyCommonTarget.valueAdapter;
-            newTracks.push(track);
-            return track;
-        });
-
-        for (const legacyCurve of legacyCurves) {
-            const legacyCurveData = legacyCurve.data;
-            const legacyValues = legacyCurveData.values;
-            if (legacyValues.length === 0) {
-                // Legacy clip did not record type info.
-                continue;
-            }
-            const legacyKeysIndex = legacyCurveData.keys;
-            // Rule: negative index means single frame.
-            const times = legacyKeysIndex < 0 ? [0.0] : legacyKeys[legacyCurveData.keys];
-            const firstValue = legacyValues[0];
-            // Rule: default to true.
-            const interpolate = legacyCurveData ?? true;
-            // Rule: _arrayLength only used for morph target, internally.
-            assertIsTrue(typeof legacyCurveData._arrayLength !== 'number' || typeof firstValue === 'number');
-            const legacyEasingMethodConverter = new legacy.LegacyEasingMethodConverter(legacyCurveData, times.length);
-
-            const installPathAndSetter = (track: Track) => {
-                track.path = legacyCurve.modifiers;
-                track.setter = legacyCurve.valueAdapter;
-            };
-
-            let legacyCommonTargetCurve: RealCurve | undefined;
-            if (typeof legacyCurve.commonTarget === 'number') {
-                // Rule: common targets should only target Vectors/`Size`/`Color`.
-                if (!legacyValues.every((value) => typeof value === 'number')) {
-                    warn(`Incorrect curve.`);
-                    continue;
-                }
-                // Rule: Each curve that has common target should be numeric curve and targets string property.
-                if (legacyCurve.valueAdapter || legacyCurve.modifiers.length !== 1 || typeof legacyCurve.modifiers[0] !== 'string') {
-                    warn(`Incorrect curve.`);
-                    continue;
-                }
-                const propertyName = legacyCurve.modifiers[0];
-                const untypedTrack = untypedTracks[legacyCurve.commonTarget];
-                const { curve } = untypedTrack.addChannel(propertyName);
-                legacyCommonTargetCurve = curve;
-            }
-
-            const convertCurve = () => {
-                if (typeof firstValue === 'number') {
-                    if (!legacyValues.every((value) => typeof value === 'number')) {
-                        warn(`Misconfigured curve.`);
-                        return;
-                    }
-                    let realCurve: RealCurve;
-                    if (legacyCommonTargetCurve) {
-                        realCurve = legacyCommonTargetCurve;
-                    } else {
-                        const track = new RealTrack();
-                        installPathAndSetter(track);
-                        newTracks.push(track);
-                        realCurve = track.channel.curve;
-                    }
-                    const interpMethod = interpolate ? RealInterpMode.LINEAR : RealInterpMode.CONSTANT;
-                    realCurve.assignSorted(times, (legacyValues as number[]).map((value) => new RealKeyframeValue({ value, interpMode: interpMethod })));
-                    legacyEasingMethodConverter.convert(realCurve);
-                    return;
-                } else if (typeof firstValue === 'object') {
-                    switch (true) {
-                    default:
-                        break;
-                    case legacyValues.every((value) => value instanceof Vec2):
-                    case legacyValues.every((value) => value instanceof Vec3):
-                    case legacyValues.every((value) => value instanceof Vec4): {
-                        type Vec4plus = Vec4[];
-                        type Vec3plus = (Vec3 | Vec4)[];
-                        type Vec2plus = (Vec2 | Vec3 | Vec4)[];
-                        const components = firstValue instanceof Vec2 ? 2 : firstValue instanceof Vec3 ? 3 : 4;
-                        const track = new VectorTrack();
-                        installPathAndSetter(track);
-                        track.componentsCount = components;
-                        const [{ curve: x }, { curve: y }, { curve: z }, { curve: w }] = track.getChannels();
-                        const interpMethod = interpolate ? RealInterpMode.LINEAR : RealInterpMode.CONSTANT;
-                        const valueToFrame = (value: number): RealKeyframeValue => new RealKeyframeValue({ value, interpMode: interpMethod });
-                        switch (components) {
-                        case 4:
-                            w.assignSorted(times, (legacyValues as Vec4plus).map((value) => valueToFrame(value.w)));
-                            legacyEasingMethodConverter.convert(w);
-                            // falls through
-                        case 3:
-                            z.assignSorted(times, (legacyValues as Vec3plus).map((value) => valueToFrame(value.z)));
-                            legacyEasingMethodConverter.convert(z);
-                            // falls through
-                        default:
-                            x.assignSorted(times, (legacyValues as Vec2plus).map((value) => valueToFrame(value.x)));
-                            legacyEasingMethodConverter.convert(x);
-                            y.assignSorted(times, (legacyValues as Vec2plus).map((value) => valueToFrame(value.y)));
-                            legacyEasingMethodConverter.convert(y);
-                            break;
-                        }
-                        newTracks.push(track);
-                        return;
-                    }
-                    case legacyValues.every((value) => value instanceof Quat): {
-                        assertIsTrue(legacyEasingMethodConverter.nil);
-                        const track = new QuaternionTrack();
-                        installPathAndSetter(track);
-                        const interpMode = interpolate ? QuaternionInterpMode.SLERP : QuaternionInterpMode.CONSTANT;
-                        track.channel.curve.assignSorted(times, (legacyValues as Quat[]).map((value) => ({
-                            value: Quat.clone(value),
-                            interpMode,
-                        })));
-                        newTracks.push(track);
-                        return;
-                    }
-                    case legacyValues.every((value) => value instanceof Color): {
-                        const track = new ColorTrack();
-                        installPathAndSetter(track);
-                        const [{ curve: r }, { curve: g }, { curve: b }, { curve: a }] = track.getChannels();
-                        const interpMethod = interpolate ? RealInterpMode.LINEAR : RealInterpMode.CONSTANT;
-                        const valueToFrame = (value: number): RealKeyframeValue => new RealKeyframeValue({ value, interpMode: interpMethod });
-                        r.assignSorted(times, (legacyValues as Color[]).map((value) => valueToFrame(value.r)));
-                        legacyEasingMethodConverter.convert(r);
-                        g.assignSorted(times, (legacyValues as Color[]).map((value) => valueToFrame(value.g)));
-                        legacyEasingMethodConverter.convert(g);
-                        b.assignSorted(times, (legacyValues as Color[]).map((value) => valueToFrame(value.b)));
-                        legacyEasingMethodConverter.convert(b);
-                        a.assignSorted(times, (legacyValues as Color[]).map((value) => valueToFrame(value.a)));
-                        legacyEasingMethodConverter.convert(a);
-                        newTracks.push(track);
-                        return;
-                    }
-                    case legacyValues.every((value) => value instanceof CubicSplineNumberValue): {
-                        assertIsTrue(legacyEasingMethodConverter.nil);
-                        const track = new RealTrack();
-                        installPathAndSetter(track);
-                        const interpMethod = interpolate ? RealInterpMode.CUBIC : RealInterpMode.CONSTANT;
-                        track.channel.curve.assignSorted(times, (legacyValues as CubicSplineNumberValue[]).map((value) => new RealKeyframeValue({
-                            value: value.dataPoint,
-                            startTangent: value.inTangent,
-                            endTangent: value.outTangent,
-                            interpMode: interpMethod,
-                        })));
-                        newTracks.push(track);
-                        return;
-                    }
-                    case legacyValues.every((value) => value instanceof CubicSplineVec2Value):
-                    case legacyValues.every((value) => value instanceof CubicSplineVec3Value):
-                    case legacyValues.every((value) => value instanceof CubicSplineVec4Value): {
-                        assertIsTrue(legacyEasingMethodConverter.nil);
-                        type Vec4plus = CubicSplineVec4Value[];
-                        type Vec3plus = (CubicSplineVec3Value | CubicSplineVec4Value)[];
-                        type Vec2plus = (CubicSplineVec2Value | CubicSplineVec3Value | CubicSplineVec4Value)[];
-                        const components = firstValue instanceof CubicSplineVec2Value ? 2 : firstValue instanceof CubicSplineVec3Value ? 3 : 4;
-                        const track = new VectorTrack();
-                        installPathAndSetter(track);
-                        track.componentsCount = components;
-                        const [x, y, z, w] = track.getChannels();
-                        const interpMethod = interpolate ? RealInterpMode.LINEAR : RealInterpMode.CONSTANT;
-                        const valueToFrame = (value: number, startTangent: number, endTangent: number): RealKeyframeValue => new RealKeyframeValue({
-                            value,
-                            startTangent,
-                            endTangent,
-                            interpMode: interpMethod,
-                        });
-                        switch (components) {
-                        case 4:
-                            w.curve.assignSorted(times, (legacyValues as Vec4plus).map(
-                                (value) => valueToFrame(value.dataPoint.w, value.inTangent.w, value.outTangent.w),
-                            ));
-                            // falls through
-                        case 3:
-                            z.curve.assignSorted(times, (legacyValues as Vec3plus).map(
-                                (value) => valueToFrame(value.dataPoint.z, value.inTangent.z, value.outTangent.z),
-                            ));
-                            // falls through
-                        default:
-                            x.curve.assignSorted(times, (legacyValues as Vec2plus).map(
-                                (value) => valueToFrame(value.dataPoint.y, value.inTangent.y, value.outTangent.y),
-                            ));
-                            y.curve.assignSorted(times, (legacyValues as Vec2plus).map(
-                                (value) => valueToFrame(value.dataPoint.x, value.inTangent.x, value.outTangent.x),
-                            ));
-                            break;
-                        }
-                        newTracks.push(track);
-                        return;
-                    }
-                    } // End switch
-                }
-
-                const objectTrack = new ObjectTrack();
-                installPathAndSetter(objectTrack);
-                objectTrack.channel.curve.assignSorted(times, legacyValues);
-                newTracks.push(objectTrack);
-            };
-
-            convertCurve();
-        }
-
+        const newTracks = legacy.convertAnimationClipLegacyData(legacyData);
         for (const track of newTracks) {
             this.addTrack(track);
         }
@@ -1477,332 +803,11 @@ export class AnimationClip extends Asset {
 
 legacyCC.AnimationClip = AnimationClip;
 
-type Binder = (path: TrackPath, setter: IValueProxyFactory | undefined) => undefined | RuntimeBinding;
-
-type RuntimeBinding = {
-    setValue(value: unknown): void;
-
-    getValue?(): unknown;
-};
-
 // #region Data compression
-
-@ccclass(`${CLASS_NAME_PREFIX_ANIM}CompressedData`)
-class CompressedData {
-    public compressRealTrack (track: RealTrack) {
-        const curve = track.channel.curve;
-        const mayBeCompressed = KeySharedRealCurves.allowedForCurve(curve);
-        if (!mayBeCompressed) {
-            return false;
-        }
-        this._tracks.push({
-            type: CompressedDataTrackType.FLOAT,
-            path: track.path,
-            setter: track.setter,
-            components: [this._addRealCurve(curve)],
-        });
-        return true;
-    }
-
-    public compressVectorTrack (vectorTrack: VectorTrack) {
-        const nComponents = vectorTrack.componentsCount;
-        const channels = vectorTrack.getChannels();
-        const mayBeCompressed = channels.every(({ curve }) => KeySharedRealCurves.allowedForCurve(curve));
-        if (!mayBeCompressed) {
-            return false;
-        }
-        const components = new Array<CompressedCurvePointer>(nComponents);
-        for (let i = 0; i < nComponents; ++i) {
-            const channel = channels[i];
-            components[i] = this._addRealCurve(channel.curve);
-        }
-        this._tracks.push({
-            type:
-                nComponents === 2
-                    ? CompressedDataTrackType.VEC2
-                    : nComponents === 3
-                        ? CompressedDataTrackType.VEC3
-                        : CompressedDataTrackType.VEC4,
-            path: vectorTrack.path,
-            setter: vectorTrack.setter,
-            components,
-        });
-        return true;
-    }
-
-    public compressQuatTrack (track: QuaternionTrack) {
-        const curve = track.channel.curve;
-        const mayBeCompressed = KeySharedQuaternionCurves.allowedForCurve(curve);
-        if (!mayBeCompressed) {
-            return false;
-        }
-        this._quatTracks.push({
-            path: track.path,
-            setter: track.setter,
-            pointer: this._addQuaternionCurve(curve),
-        });
-        return true;
-    }
-
-    public createEval (binder: Binder) {
-        const compressedDataEvalStatus: CompressedDataEvalStatus = {
-            keySharedCurvesEvalStatuses: [],
-            trackEvalStatuses: [],
-            keysSharedQuatCurvesEvalStatues: [],
-            quatTrackEvalStatuses: [],
-        };
-
-        const {
-            keySharedCurvesEvalStatuses,
-            trackEvalStatuses,
-            keysSharedQuatCurvesEvalStatues,
-            quatTrackEvalStatuses,
-        } = compressedDataEvalStatus;
-
-        for (const curves of this._curves) {
-            keySharedCurvesEvalStatuses.push({
-                curves,
-                result: new Array(curves.curveCount).fill(0.0),
-            });
-        }
-
-        for (const track of this._tracks) {
-            const trackTarget = binder(track.path, track.setter);
-            if (!trackTarget) {
-                continue;
-            }
-            let immediate: CompressedTrackImmediate | undefined;
-            switch (track.type) {
-            default:
-            case CompressedDataTrackType.FLOAT:
-                break;
-            case CompressedDataTrackType.VEC2:
-                immediate = new Vec2();
-                break;
-            case CompressedDataTrackType.VEC3:
-                immediate = new Vec3();
-                break;
-            case CompressedDataTrackType.VEC4:
-                immediate = new Vec4();
-                break;
-            }
-            trackEvalStatuses.push({
-                type: track.type,
-                target: trackTarget,
-                curves: track.components,
-                immediate,
-            });
-        }
-
-        for (const curves of this._quatCurves) {
-            keysSharedQuatCurvesEvalStatues.push({
-                curves,
-                result: Array.from({ length: curves.curveCount }, () => new Quat()),
-            });
-        }
-
-        for (const track of this._quatTracks) {
-            const trackTarget = binder(track.path, track.setter);
-            if (!trackTarget) {
-                continue;
-            }
-            quatTrackEvalStatuses.push({
-                target: trackTarget,
-                curve: track.pointer,
-            });
-        }
-
-        return new CompressedDataEvaluator(compressedDataEvalStatus);
-    }
-
-    public collectAnimatedJoints () {
-        const joints: string[] = [];
-
-        for (const track of this._tracks) {
-            if (!track.setter && isTargetingTRS(track.path)) {
-                const { path } = track.path[0];
-                joints.push(path);
-            }
-        }
-
-        return joints;
-    }
-
-    @serializable
-    private _curves: KeySharedRealCurves[] = [];
-
-    @serializable
-    private _tracks: CompressedTrack[] = [];
-
-    @serializable
-    private _quatCurves: KeySharedQuaternionCurves[] = [];
-
-    @serializable
-    private _quatTracks: CompressedQuatTrack[] = [];
-
-    private _addRealCurve (curve: RealCurve): CompressedCurvePointer {
-        const times = Array.from(curve.times());
-        let iKeySharedCurves = this._curves.findIndex((shared) => shared.matchCurve(curve));
-        if (iKeySharedCurves < 0) {
-            iKeySharedCurves = this._curves.length;
-            const keySharedCurves = new KeySharedRealCurves(times);
-            this._curves.push(keySharedCurves);
-        }
-        const iCurve = this._curves[iKeySharedCurves].curveCount;
-        this._curves[iKeySharedCurves].addCurve(curve);
-        return {
-            shared: iKeySharedCurves,
-            component: iCurve,
-        };
-    }
-
-    public _addQuaternionCurve (curve: QuaternionCurve): CompressedQuatCurvePointer {
-        const times = Array.from(curve.times());
-        let iKeySharedCurves = this._quatCurves.findIndex((shared) => shared.matchCurve(curve));
-        if (iKeySharedCurves < 0) {
-            iKeySharedCurves = this._quatCurves.length;
-            const keySharedCurves = new KeySharedQuaternionCurves(times);
-            this._quatCurves.push(keySharedCurves);
-        }
-        const iCurve = this._quatCurves[iKeySharedCurves].curveCount;
-        this._quatCurves[iKeySharedCurves].addCurve(curve);
-        return {
-            shared: iKeySharedCurves,
-            curve: iCurve,
-        };
-    }
-
-    public validate () {
-        return this._tracks.length > 0;
-    }
-}
-
-class CompressedDataEvaluator {
-    constructor (compressedDataEvalStatus: CompressedDataEvalStatus) {
-        this._compressedDataEvalStatus = compressedDataEvalStatus;
-    }
-
-    public evaluate (time: number) {
-        const {
-            keySharedCurvesEvalStatuses,
-            trackEvalStatuses: compressedTrackEvalStatuses,
-            keysSharedQuatCurvesEvalStatues,
-            quatTrackEvalStatuses,
-        } = this._compressedDataEvalStatus;
-
-        const getPreEvaluated = (pointer: CompressedCurvePointer) => keySharedCurvesEvalStatuses[pointer.shared].result[pointer.component];
-
-        for (const { curves, result } of keySharedCurvesEvalStatuses) {
-            curves.evaluate(time, result);
-        }
-
-        for (const { type, target, immediate, curves } of compressedTrackEvalStatuses) {
-            let value: unknown = immediate;
-            switch (type) {
-            default:
-                break;
-            case CompressedDataTrackType.FLOAT:
-                value = getPreEvaluated(curves[0]);
-                break;
-            case CompressedDataTrackType.VEC2:
-                Vec2.set(
-                    value as Vec2,
-                    getPreEvaluated(curves[0]),
-                    getPreEvaluated(curves[1]),
-                );
-                break;
-            case CompressedDataTrackType.VEC3:
-                Vec3.set(
-                    value as Vec3,
-                    getPreEvaluated(curves[0]),
-                    getPreEvaluated(curves[1]),
-                    getPreEvaluated(curves[2]),
-                );
-                break;
-            case CompressedDataTrackType.VEC4:
-                Vec4.set(
-                    value as Vec4,
-                    getPreEvaluated(curves[0]),
-                    getPreEvaluated(curves[1]),
-                    getPreEvaluated(curves[2]),
-                    getPreEvaluated(curves[4]),
-                );
-                break;
-            }
-            target.setValue(value);
-        }
-
-        for (const { curves, result } of keysSharedQuatCurvesEvalStatues) {
-            curves.evaluate(time, result);
-        }
-
-        for (const { target, curve } of quatTrackEvalStatuses) {
-            target.setValue(keysSharedQuatCurvesEvalStatues[curve.shared].result[curve.curve]);
-        }
-    }
-
-    private _compressedDataEvalStatus: CompressedDataEvalStatus;
-}
-
-interface CompressedTrack {
-    path: TrackPath;
-    setter: IValueProxyFactory | undefined;
-    type: CompressedDataTrackType;
-    components: CompressedCurvePointer[];
-}
-
-interface CompressedQuatTrack {
-    path: TrackPath;
-    setter: IValueProxyFactory | undefined;
-    pointer: CompressedQuatCurvePointer;
-}
-
-enum CompressedDataTrackType {
-    FLOAT,
-    VEC2,
-    VEC3,
-    VEC4,
-}
 
 interface TrackEvalStatus {
     binding: RuntimeBinding;
     trackEval: TrackEval;
-}
-
-type CompressedTrackImmediate = Vec2 | Vec3 | Vec4;
-
-interface CompressedDataEvalStatus {
-    keySharedCurvesEvalStatuses: Array<{
-        curves: KeySharedRealCurves;
-        result: number[];
-    }>;
-
-    trackEvalStatuses: Array<{
-        type: CompressedDataTrackType;
-        target: RuntimeBinding;
-        immediate: CompressedTrackImmediate | undefined;
-        curves: CompressedCurvePointer[];
-    }>;
-
-    keysSharedQuatCurvesEvalStatues: Array<{
-        curves: KeySharedQuaternionCurves;
-        result: Quat[];
-    }>;
-
-    quatTrackEvalStatuses: Array<{
-        target: RuntimeBinding;
-        curve: CompressedQuatCurvePointer;
-    }>;
-}
-
-interface CompressedCurvePointer {
-    shared: number;
-    component: number;
-}
-
-interface CompressedQuatCurvePointer {
-    shared: number;
-    curve: number;
 }
 
 // #endregion
@@ -2067,31 +1072,6 @@ function createGeneralBinding (
         }
     }
     return null;
-}
-
-type TrsTrackPath = [HierarchyPath, 'position' | 'rotation' | 'scale' | 'eulerAngles'];
-
-function isTargetingTRS (path: TargetPath[]): path is TrsTrackPath {
-    let prs: string | undefined;
-    if (path.length === 1 && typeof path[0] === 'string') {
-        prs = path[0];
-    } else if (path.length > 1) {
-        for (let i = 0; i < path.length - 1; ++i) {
-            if (!(path[i] instanceof HierarchyPath)) {
-                return false;
-            }
-        }
-        prs = path[path.length - 1] as string;
-    }
-    switch (prs) {
-    case 'position':
-    case 'scale':
-    case 'rotation':
-    case 'eulerAngles':
-        return true;
-    default:
-        return false;
-    }
 }
 
 function createRuntimeBinding (target: unknown, trackPath: TrackPath, setter?: IValueProxyFactory): null | RuntimeBinding {
