+++ cocos/core/animation/define.ts
@@ -1,0 +1,3 @@
+export const CLASS_NAME_PREFIX_ANIM = 'cc.animation.';
+
+export const createEvalSymbol = Symbol('CreateEval');
