+++ cocos/core/animation/tracks/object-track.ts
@@ -1,0 +1,11 @@
+import { ccclass } from 'cc.decorator';
+import { ObjectCurve } from '../../curves';
+import { CLASS_NAME_PREFIX_ANIM, createEvalSymbol } from '../define';
+import { SingleChannelTrack } from './track';
+
+@ccclass(`${CLASS_NAME_PREFIX_ANIM}ObjectTrack`)
+export class ObjectTrack<T> extends SingleChannelTrack<ObjectCurve<unknown>> {
+    protected createCurve () {
+        return new ObjectCurve<T>();
+    }
+}
