+++ cocos/core/animation/tracks/utils.ts
@@ -1,0 +1,10 @@
+import type { Curve } from './track';
+
+export function maskIfEmpty<T extends Curve> (curve: T) {
+    return curve.empty ? undefined : curve;
+}
+
+export interface Range {
+    min: number;
+    max: number;
+}
